Correcting Chris's spelling.
He's not The Massigh, he's a very naughty boy.

+++ HumanitarianApp/app/src/main/java/edu/rose_hulman/srproject/humanitarianapp/serialisation/SerilizationConstants.java
@@ -21,7 +21,7 @@ public class SerilizationConstants {
 
     public static long generateID(long objectEnum) {
         long value = objectEnum << TYPE_LOCATION;
-        value = value | ((long) ApplicationWideData.userID) << CREATER_LOCATION;
+        value = value | ApplicationWideData.userID << CREATER_LOCATION;
         value = value | ((long) ApplicationWideData.createdObjectCounter) << COUNTER_LOCATION;
         ApplicationWideData.createdObjectCounter ++;
         return value;
